Corrections for header_service_www_output
The variable headerservice::header_service_www_output was changed from an Array to an String, and the puppet module deals with checking the full path.

Lines 106 to 122 deal with the path verification of the aforementioned variable.

The setup_sal.env file was replaced by a variable and it was added as a requirement.

filepath option for the headerservice resouce was changed to use the right variable.